410/514

+++ ft_printf.h
@@ -56,7 +56,7 @@ void				ft_intzap(char *tstr, t_list *list, int *i, int *j);
 void				ft_intzappoz(char *tstr, t_list *list, int *i, int *j);
 char				*ft_longtoa(long int number);
 int					ft_uintcheck(long int temp, t_list *list);
-int					ft_hexproc(t_list *list, long int temp, int flag);
+int					ft_hexproc(t_list *list, unsigned int temp, int flag);
 void				ft_hexwidthpon(t_list *list, int len);
 void				ft_hexprint(t_list *list, char *tstr);
 int					ft_atoi(const char *str);
